I looked at buckets recently and noticed that especially after recent refactoring random lookups for `y` needed during both tables creation and proving are actually happening in sync with reading corresponding positions. Since reading positions without reading `y`s is not really happening, why not storing both together?

Moreover, storing full `y` is not actually necessary, we could store just `r` and do cheap addition based on a bucket value is in, but it ended up being a tiny bit slower from my testing (but it would use a bit less memory).

With a sprinkle of cache alignment for `LeftTargets` we have another major performance improvement and I think that actually sorting revisit might be happening soon, but this time using a custom radix sort implementation.

Results are still a bit flaky, but the minimums are in fact moving down:
```
Before:
chia/table/single/1x    time:   [702.17 ms 707.33 ms 712.59 ms]
                        thrpt:  [1.4033  elem/s 1.4138  elem/s 1.4242  elem/s]
chia/table/parallel/8x  time:   [502.50 ms 505.77 ms 509.35 ms]
                        thrpt:  [15.706  elem/s 15.818  elem/s 15.920  elem/s]
chia/proof/missing      time:   [107.29 ns 108.80 ns 111.58 ns]
                        thrpt:  [8.9618 Melem/s 9.1915 Melem/s 9.3207 Melem/s]
chia/proof/present      time:   [374.03 ns 375.88 ns 377.16 ns]
                        thrpt:  [2.6514 Melem/s 2.6604 Melem/s 2.6736 Melem/s]
After:
chia/table/single/1x    time:   [638.21 ms 642.97 ms 649.77 ms]
                        thrpt:  [1.5390  elem/s 1.5553  elem/s 1.5669  elem/s]
chia/table/parallel/8x  time:   [426.94 ms 431.49 ms 436.70 ms]
                        thrpt:  [18.319  elem/s 18.540  elem/s 18.738  elem/s]
chia/proof/missing      time:   [76.806 ns 76.913 ns 77.038 ns]
                        thrpt:  [12.981 Melem/s 13.002 Melem/s 13.020 Melem/s]
chia/proof/present      time:   [351.19 ns 352.22 ns 354.05 ns]
                        thrpt:  [2.8245 Melem/s 2.8391 Melem/s 2.8474 Melem/s]
```

Overall, performance more than doubled since May when the last (as of right now) Subspace backport has happened. And as a bonus, this PR actually removes more code than adds, which is not always the case.

I'll check whether storing metadata next to positions and `y` is worth it (probably not) separately.